+++ react/components/ToggleLayout.ts
@@ -1,0 +1,3 @@
+import ToggleLayout from './ToggleLayout/ToggleLayout';
+
+export default ToggleLayout;