+++ react/components/ToggleLayout/ToggleLayout.types.ts
@@ -1,0 +1,6 @@
+interface ToggleLayoutProps {
+  enabled: boolean;
+  children: React.ReactNode;
+}
+
+export default ToggleLayoutProps;