+++ react/components/ToggleLayout/index.ts
@@ -1,0 +1,3 @@
+import ToggleLayout from './ToggleLayout';
+
+export default ToggleLayout;