+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Added new `ToggleLayout` component.
 - Added new `Iframe` component.
 
 ## [1.0.0]